requirements.txt
Django==5.0.7
psycopg[binary]==3.2.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
PyJWT==2.9.0
bcrypt==4.1.3
python-dotenv==1.0.1


pytest==8.3.2
pytest-django==4.9.0


black==24.8.0
isort==5.13.2
ruff==0.6.2
pre-commit==3.8.0